Status bar: memory display with brightness.
This commit implements the memory display where the icon glows more
depending on the usage and a usage label in gigabyte is right next to
the icon. Later commits might darken the icon when the usage is not high
and moving particles should be shown when the usage is very high.

This commit also added the BlockOverlay function and temporarily added
assertion to object unpacking.